// types/enum/EDevice.enum.ts
export enum EDevice {
  SENSOR = "SENSOR",
  AKTUATOR = "AKTUATOR",
  SENSOR_TEMPERATURE = "SENSOR_TEMPERATURE",
  SENSOR_WATER_LEVEL = "SENSOR_WATER_LEVEL",
  SENSOR_PARKING = "SENSOR_PARKING",
}

export enum EDeviceStatus {
  ON = "ON",
  OFF = "OFF",
}

export enum EStatus {
  ON = "1",
  OFF = "0",
}